fix bug
in c++, we can’t call subclass method in base class. In our case,
when initiate a subclass of persistent_object, base class constructor actually
 didn’t call subclass `class_name`, but base class’s.
To solve this, persistent_object won’t get the class name automatically
from its actual class(it just can’t), but use an explicit argument of the
constructor.

And also, it is the persistent_object’s responsibility to make SQLs, not
the hyper_table’s.