[unperfect]fetch correspond object in another class with 1-to-1 relation.
If one column of a class defined as some other class in configuration
file, as a one-to-one relation, we’ll fetch the mapped object as the value
as the primary key value as well. Though, it mapped as a share_ptr of
persistent_object, which is the reason I name it unperfectly.

To achieve this, interface of Connection::query changed from accepting
a reference to accepting a share_ptr. As experiment:
    vector<shared_ptr<Base>> vec;
    vec.push_back(make_shared<Derive>(new Derive());
    vec.front()->do_something();

would do the Derive class operation, so we just specify the type as
persistent_object, no longer using a template argument. This results the
 registered as a unordered_map of string->shared_ptr<persistent_object>.

What’s more, persistent_object has one more member variable my_class_name.
Since creating mapped object in query just return persistent_object, the
class_name information would lost.